Step update
move data of step 6 to step 3
